+++ src/server/socket-proxy.ts
@@ -1,10 +1,10 @@
-import fs from 'fs-extra'
 import { Server as SocketServer } from 'socket.io'
 import { io as ClientIO } from 'socket.io-client'
 import { HttpServer, ViteDevServer } from 'vite'
 import { context, ENVOptions } from 'src/context'
 import { handlebarsTracker } from 'src/server/trackers/handlebars-tracker'
 import pathUtils from 'src/utils/path-utils'
+import FsUtils from 'src/utils/fs-utils'
 
 export default function socketProxy(server: ViteDevServer) {
   const env = context.env as ENVOptions
@@ -18,13 +18,14 @@ export default function socketProxy(server: ViteDevServer) {
     })
 
     // Browser >>> Foundry [intercept templating calls]
-    socket.onAny((event, ...args) => {
+    socket.onAny(async (event, ...args) => {
       const maybeAck = typeof args[args.length - 1] === 'function' ? args.pop() : null
 
       if (event === 'template') {
-        const localPath = pathUtils.foundryVTTUrlToLocal(args[0])
+        const localPath = await pathUtils.foundryVTTUrlToLocal(args[0])
         if (localPath) {
-          if (maybeAck) maybeAck({ html: fs.readFileSync(localPath, 'utf8'), success: true })
+          const html = await FsUtils.readFile(localPath)
+          if (maybeAck) maybeAck({ html, success: true })
           handlebarsTracker.addFile(args[0], localPath)
           return
         }